move diagnostic scripts to own file
