+++ tests/integration/test_diagnostic_run.py
@@ -1,6 +1,8 @@
 """Test diagnostic script runs."""
 import contextlib
+import shutil
 import sys
+from pathlib import Path
 from textwrap import dedent
 
 import pytest
@@ -49,81 +51,34 @@ def check(result_file):
     assert not missing
 
 
-RECIPE_PY = """
-import yaml
-from esmvaltool.diag_scripts.shared import run_diagnostic
-
-def main(cfg):
-    with open(cfg['setting_name'], 'w') as file:
-        yaml.safe_dump(cfg, file)
-
-if __name__ == '__main__':
-    with run_diagnostic() as config:
-        main(config)
-"""
-RECIPE_NCL = """
-begin
-    print("INFO    Loading settings from " + getenv("settings"))
-    loadscript("$settings")
-end
-print("INFO Writing " + diag_script_info@setting_name)
-n = str_get_nl()
-result = "run_dir: " + config_user_info@run_dir + n +\
-            "work_dir: " + config_user_info@work_dir + n +\
-            "plot_dir: " + config_user_info@plot_dir + n +\
-            "log_level: " + config_user_info@log_level + n +\
-            "input_files: []" + n
-
-system("echo '" + result + "' > " + diag_script_info@setting_name)
-"""
-RECIPE_R = """
-library(yaml)
-args <- commandArgs(trailingOnly = TRUE)
-print(paste0("INFO    Loading settings from ", args[1]))
-settings <- yaml::read_yaml(args[1])
-
-print(paste0("INFO    Writing settings to ", settings$setting_name))
-yaml::write_yaml(settings, settings$setting_name)
-"""
-RECIPE_JL = """
-import YAML
-@info "Starting diagnostic script with" ARGS
-config_file = ARGS[1]
-cfg = YAML.load_file(config_file)
-out_file = cfg["setting_name"]
-@info "Copying file to" out_file
-Base.Filesystem.cp(config_file, out_file)
-@info "Done"
-"""
-
-SCRIPTS = [('diagnostic.py', RECIPE_PY),
-           pytest.param('diagnostic.ncl',
-                        RECIPE_NCL,
-                        marks=pytest.mark.skipif(
-                            sys.platform == 'darwin',
-                            reason="ESMValTool ncl not supported on OSX")),
-           pytest.param('diagnostic.R',
-                        RECIPE_R,
-                        marks=pytest.mark.skipif(
-                            sys.platform == 'darwin',
-                            reason="ESMValTool R not supported on OSX")),
-           pytest.param('diagnostic.jl',
-                        RECIPE_JL,
-                        marks=pytest.mark.skipif(
-                            sys.platform == 'darwin',
-                            reason="ESMValTool Julia not supported on OSX"))]
+SCRIPTS = [
+    'diagnostic.py',
+    pytest.param('diagnostic.ncl',
+                 marks=pytest.mark.skipif(
+                     sys.platform == 'darwin',
+                     reason="ESMValTool ncl not supported on OSX")),
+    pytest.param('diagnostic.R',
+                 marks=pytest.mark.skipif(
+                     sys.platform == 'darwin',
+                     reason="ESMValTool R not supported on OSX")),
+    pytest.param('diagnostic.jl',
+                 marks=pytest.mark.skipif(
+                     sys.platform == 'darwin',
+                     reason="ESMValTool Julia not supported on OSX"))
+]
 
 
 @pytest.mark.installation
-@pytest.mark.parametrize('script_file, script', SCRIPTS)
-def test_diagnostic_run(tmp_path, script_file, script):
+@pytest.mark.parametrize('script_file', SCRIPTS)
+def test_diagnostic_run(tmp_path, script_file):
+
+    local_script_file = Path(__file__).parent / script_file
 
     recipe_file = tmp_path / 'recipe_test.yml'
     script_file = tmp_path / script_file
     result_file = tmp_path / 'result.yml'
 
-    # Write script to file
-    script_file.write_text(str(script))
+    shutil.copy(local_script_file, script_file)
 
     # Create recipe
     recipe = dedent("""
